# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TripUAI API",
        "description": "Documentação da API para a Aplicação TripUAI",
        "contact": {
            "email": "admin@tripuai.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Login do usuário e retorno do token JWT",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@tripuai.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Obter perfil do usuário logado",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "Dados do perfil do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Admin User"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "admin@tripuai.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Logout do usuário",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/broadcast": {
            "post": {
                "tags": [
                    "Administrativo"
                ],
                "summary": "Enviar email para todos os usuários",
                "operationId": "e4a89824c99aa6b874de315e95f09e4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subject",
                                    "body"
                                ],
                                "properties": {
                                    "subject": {
                                        "type": "string",
                                        "example": "Nova Funcionalidade"
                                    },
                                    "body": {
                                        "type": "string",
                                        "example": "Nós lançamos uma nova funcionalidade..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Processamento de envio iniciado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O processamento do broadcast de email foi iniciado em segundo plano."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests": {
            "get": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Listar solicitações de viagem",
                "operationId": "cc8f5935116993c7bce47473cf6e44f1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "solicitado",
                                "aprovado",
                                "cancelado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de solicitações de viagem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Criar uma nova solicitação de viagem",
                "operationId": "012a77a49c7f0bc54b6602e659215010",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "requester_name",
                                    "destination",
                                    "departure_date",
                                    "return_date"
                                ],
                                "properties": {
                                    "requester_name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-10"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitação de viagem criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Viagem solicitada com sucesso!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{travelRequest}/approve": {
            "patch": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Aprovar uma solicitação de viagem (Apenas admin)",
                "operationId": "a99c40b34bf3f3ba89342174fa27a569",
                "parameters": [
                    {
                        "name": "travelRequest",
                        "in": "path",
                        "description": "ID da solicitação de viagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitação de viagem aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Viagem aprovada com sucesso!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{travelRequest}/cancel": {
            "patch": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Cancelar uma solicitação de viagem",
                "operationId": "9d992549fb060c5cd69bc3a44b71f3ea",
                "parameters": [
                    {
                        "name": "travelRequest",
                        "in": "path",
                        "description": "ID da solicitação de viagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitação de viagem cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Viagem cancelada."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação (ex: não pode cancelar solicitação aprovada)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/travel-requests/{travelRequest}/show": {
            "get": {
                "tags": [
                    "Solicitações de Viagem"
                ],
                "summary": "Obter detalhes de uma solicitação de viagem",
                "operationId": "612276ba7fcde0ce0a016f74464aa973",
                "parameters": [
                    {
                        "name": "travelRequest",
                        "in": "path",
                        "description": "ID da solicitação de viagem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da solicitação de viagem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entre com o token JWT Bearer",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "Administrativo",
            "description": "Administrativo"
        },
        {
            "name": "Solicitações de Viagem",
            "description": "Solicitações de Viagem"
        }
    ]
}